requirements.txt
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.4
django-cors-headers==2.1.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
pkg-resources==0.0.0
pyinotify==0.9.6
PyJWT==1.5.3
python-termstyle==0.1.10
pytz==2017.2
requests==2.18.4
sniffer==0.4.0
uritemplate==3.0.0
urllib3==1.22
